Main library file
v1.0 of the graphslib library for Python3. Ready with weighted and not graph classes, binary indexed tree class, node and vertex classes. Apropriate algorithms for parsing are also included.
MST: Kruskal's algorithm on WeightedGraph
Shortest Path: Dijkstra's algorithm on WeightedGraph
Sorting: Binary indexed tree algorithm in BinaryIndexedTree

Note: this library was made in half a day to use in the IEEEXtreme13.0 competition the same night and are in no way the best implementations out there. I just needed an all together solution to use as is. Any comments, suggestions and improvements appreciated!